feat: add go.mod file to manage dependencies and specify Go version
feat: add releases.json file to store information about available releases

feat: add sync-releases.go file to synchronize releases with the releases directory

The go.mod file is added to the project to manage dependencies and specify the Go version. This will help ensure that the project can be built and run consistently across different environments.

The releases.json file is added to store information about available releases. This file will be used to keep track of the different versions of PHP that are available for download.

The sync-releases.go file is added to synchronize the releases with the releases directory. This script will read the files in the releases directory, extract the version and architecture information from the file names, and update the releases.json file accordingly. This will make it easier to manage and update the list of available releases.